+++ CHANGELOG.rst
@@ -64,6 +64,7 @@ LM PLC
 
     * fix: "out" variable is now correctly set if macro raises an exception
 
+    * LVar variable logic: normal (default) or simple
     * SSL, TLS and SMTP auth support for mailer
     * macro function "cmd": new argument "stdin_data"
     * macro function "sha256sum"